Refactor service, mapper, and dto architectur and BinanceApiService - check hidden keys
