+++ backend/src/main/java/com/tradomator/model/dto/coingeckoapi/CgeckoApiCoin.java
@@ -1,4 +1,4 @@
-package com.tradomator.model.coingeckoapi;
+package com.tradomator.model.dto.coingeckoapi;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
