+++ backend/src/main/java/com/tradomator/model/dto/coingeckoapi/ImageApi.java
@@ -1,5 +1,4 @@
-package com.tradomator.model.coingeckoapi;
-
+package com.tradomator.model.dto.coingeckoapi;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -7,12 +6,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Data
+
 @AllArgsConstructor
 @NoArgsConstructor
+@Data
 @Builder
-public class MarketDataApi {
-@JsonProperty("current_price")
-    private CurrentPriceApi currentPriceApi;
+public class ImageApi {
+    @JsonProperty("small")
+    private String imageUrl;
 
 }
