+++ backend/src/main/java/com/tradomator/model/dto/coingeckoapi/MarketDataApi.java
@@ -1,4 +1,5 @@
-package com.tradomator.model.coingeckoapi;
+package com.tradomator.model.dto.coingeckoapi;
+
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -6,13 +7,12 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
+@Data
 @AllArgsConstructor
 @NoArgsConstructor
-@Data
 @Builder
-public class ImageApi {
-    @JsonProperty("small")
-    private String imageUrl;
+public class MarketDataApi {
+@JsonProperty("current_price")
+    private CurrentPriceApi currentPriceApi;
 
 }
