+++ backend/src/main/resources/application.properties
@@ -1,3 +1,4 @@
 spring.data.mongodb.database = tmDB
 
 com.tradomator.wallet.binance.signature=${BINANCE_SIGNATURE}
+com.tradomator.wallet.binance.apikey=${BINANCE_API_KEY}
