+++ backend/src/test/java/com/tradomator/service/CoinServiceTest.java
@@ -2,12 +2,12 @@ package com.tradomator.service;
 
 import com.tradomator.model.CoinIdCard;
 import com.tradomator.model.CurrentPrice;
-import com.tradomator.model.coingeckoapi.CgeckoApiCoin;
-import com.tradomator.model.coingeckoapi.CurrentPriceApi;
-import com.tradomator.model.coingeckoapi.ImageApi;
-import com.tradomator.model.coingeckoapi.MarketDataApi;
+import com.tradomator.model.dto.coingeckoapi.CgeckoApiCoin;
+import com.tradomator.model.dto.coingeckoapi.CurrentPriceApi;
+import com.tradomator.model.dto.coingeckoapi.ImageApi;
+import com.tradomator.model.dto.coingeckoapi.MarketDataApi;
 import com.tradomator.service.coingeckoapi.CoingeckoApiService;
-import com.tradomator.service.mapperapi.CoingeckoMapper;
+import com.tradomator.mapper.CoingeckoMapper;
 
 
 import org.junit.jupiter.api.Test;
